remoteproc: ti_k3_m4_remoteproc: Fix build warnings around k3_m4_rproc_stop()
Fix below warning:

 drivers/remoteproc/ti_k3_m4_remoteproc.c: In function 'k3_m4_rproc_stop':
   drivers/remoteproc/ti_k3_m4_remoteproc.c:379:46: warning: cast to pointer from integer of different size [-Wint-to-pointer-cast]
     379 |         ret = mbox_send_message(kproc->mbox, (void *)msg);
         |                                              ^

Fixes: 99cff963c413 ("remoteproc: k3-m4: extend stop to send shutdown message to m4 core")
Reported-by: kernel test robot <lkp@intel.com>

+++ drivers/remoteproc/ti_k3_m4_remoteproc.c
@@ -364,7 +364,7 @@ static int k3_m4_rproc_stop(struct rproc *rproc)
 	unsigned long to = msecs_to_jiffies(3000);
 	struct k3_m4_rproc *kproc = rproc->priv;
 	struct device *dev = kproc->dev;
-	u32 msg = RP_MBOX_SHUTDOWN;
+	unsigned long msg = RP_MBOX_SHUTDOWN;
 	int ret;
 
 	if (kproc->ipc_only) {
